% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_filter_aux.R
\name{spat_feature}
\alias{spat_feature}
\title{Creates the description in XML format of a point, linestring or polygon}
\usage{
spat_feature(
  sptype,
  crs_in,
  coords,
  version = WFS_get_version(),
  sep = WFS_get_sep()
)
}
\arguments{
\item{sptype}{Character string with the type of spatial feature. One of \code{point}, \code{linestring} or \code{polygon} (case insensitive)}

\item{crs_in}{Character string with the Coordinate Reference System}

\item{coords}{A numeric vector of even length with the coordinates of the feature or in case of a polygon a list of
such vectors containing the outer coordinates and optionally the coordinates of holes.
Instead of a vector also a two-column matrix can be specified.}

\item{version}{Character string with the WFS request version}

\item{sep}{Character string with separator (to be used to split outer from inner polygons). Useful for printouts of query strings.}
}
\value{
Character vector with the created xml fragment
}
\description{
Can be used in \code{\link[=build_filter]{build_filter()}} or \code{\link[=spat_xml]{spat_xml()}}
}
\examples{
\dontrun{
 spat_feature('point','EPSG:28992', point_coords)
 spat_feature('Polygon','EPSG:28992',list(outer_coords,hole_coords))
}
}
