% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_filter_aux.R
\name{spat_xml}
\alias{spat_xml}
\title{Creates the spatial part of a filter in XML format}
\usage{
spat_xml(
  gemprop,
  feature,
  distance = NULL,
  version = WFS_get_version(),
  spat_fun = "DWithin",
  units = "meters"
)
}
\arguments{
\item{gemprop}{Character string with the name of the geometric field}

\item{feature}{Character with XML description of feature. Description can be result of \code{\link[=spat_feature]{spat_feature()}}}

\item{distance}{Number scalar with distance for \code{DWithin} operator. Unit given by argument \code{units}}

\item{version}{Character string with the WFS request version}

\item{spat_fun}{Character string with the name of the spatial operator}

\item{units}{Character string indicating the units of the argument \code{distance}}
}
\value{
Character vector with the created xml fragment
}
\description{
Can be used to create in XML format the spatial part of spatial operators such as
\code{Disjoint}, \code{Equals}, \code{DWithin}, \code{Beyond}, \code{Intersects}, \code{Touches}, \code{Crosses}, \code{Within},
\code{Contains}, \code{Overlaps} and \code{BBOX} .
}
\examples{
\dontrun{
spat_xml('geometrie',
      spat_feature('point','EPSG:28992', point_coords),
             50) # filters features within 59 meters of point
spat_xml('geometrie',
       spat_feature('Polygon','EPSG:28992',list(outer_coords,hole_coords)),
      spat_fun='Intersects') # filters features intersecting polygon with hole
}
}
