% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_describefeaturetype.R
\name{WFS_describefeaturetype}
\alias{WFS_describefeaturetype}
\title{DescribeFeatureType for the WFS service}
\usage{
WFS_describefeaturetype(
  typenames,
  url = WFS_get_url(),
  version = WFS_get_version(),
  debug = F,
  verbose = F,
  out_path = NULL
)
}
\arguments{
\item{typenames}{Character vector  with the name of features (such as e.g. obtained by using \code{\link[=WFS_featuretypes]{WFS_featuretypes()}} )}

\item{url}{URL of the WFS service. See \code{\link[=WFS_get_url]{WFS_get_url()}} for the default}

\item{version}{software version for WFS service request. See \code{\link[=WFS_get_version]{WFS_get_version()}} for the default}

\item{debug}{Logical indicating the httr response is to be returned}

\item{verbose}{Logical indicating full request and httr response code will be displayed}

\item{out_path}{(optional) path where the \code{FeatureType} info is to be saved in xml format. Only valid when \code{typenames} contains one featuretype}
}
\value{
a \code{tibble} with \code{typename} and \code{name} (the fieldnames)
}
\description{
Retrieve and unpack the \code{DescribeFeatureType} document for one or more  \code{FeatureTypes} into a tibble
}
\details{
\code{typenames} is copied to the result tibble to distinguish the various \code{FeatureTypes} when more than one is specified
}
\examples{
\dontrun{
WFS_describefeaturetype("topp:gidw_groenbomen")
}
}
