% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_featuretypes.R
\name{WFS_featuretypes}
\alias{WFS_featuretypes}
\title{Retrieve featuretypes information}
\usage{
WFS_featuretypes(xml_doc, filternames = NULL)
}
\arguments{
\item{xml_doc}{XML document with the output of a GetCapabilities document}

\item{filternames}{Character string that filters the returned featurenames}
}
\value{
tibble with name, default CRS, lower and upper corner (in WGS84 coordinates) for each filtered featuretype name
}
\description{
Retrieve the names of the features with spatial information from a GetCapabilities document
}
\examples{
xml_cap1 <- WFS_getcapabilities()
FeatureTypes1 <- WFS_featuretypes(xml_cap1,
       filternames=stringr::fixed("bomen", ignore_case = T))
}
