% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_getcapabilities.R
\name{WFS_getcapabilities}
\alias{WFS_getcapabilities}
\title{GetCapabilities for the WFS service}
\usage{
WFS_getcapabilities(
  url = WFS_get_url(),
  version = WFS_get_version(),
  out_path = NULL
)
}
\arguments{
\item{url}{URL of the WFS service. See \code{\link[=WFS_get_url]{WFS_get_url()}} for the default}

\item{version}{software version for WFS service request. See \code{\link[=WFS_get_version]{WFS_get_version()}} for the default}

\item{out_path}{path where the xml result is to be saved}
}
\value{
xml document with the \code{GetCapabilities} information for this WFS service
}
\description{
Retrieve the GetCapabilities document with request version \code{version} for WFS service \code{url}
}
\examples{
WFS_getcapabilities("https://geoweb.amstelveen.nl/geoserver/topp/wfs","1.1.0","gc.xml")
}
