% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_getcapabilities.R
\name{WFS_getcapabilities_POST}
\alias{WFS_getcapabilities_POST}
\title{Retrieves the GetCapabilities document}
\usage{
WFS_getcapabilities_POST(
  url = WFS_get_url(),
  version = WFS_get_version(),
  debug = F,
  httrverbose = F,
  out_path = NULL
)
}
\arguments{
\item{url}{Character with base url for WFS service}

\item{version}{software version for WFS service request. See \code{\link[=WFS_get_version]{WFS_get_version()}} for the default}

\item{debug}{Logical indicating the httr response is to be returned}

\item{out_path}{(optional) path where the xml result is to be saved}

\item{verbose}{Logical indicating full request and httr response code will be displayed}
}
\value{
xml document with the \code{GetCapabilities} information for this WFS service
or character string 'UNEXPECTED ERROR' when an error was encountered
}
\description{
Retrieves the GetCapabilities document
}
\examples{
\dontrun{
base_url  <- "https://geodata.nationaalgeoregister.nl/wijkenbuurten2019/wfs"
cap1 = WFS_getcapabilities_POST(base_url,'1.1.0')
WFS_util_showxml(cap1,lines=2:3)
cap2 <- WFS_getcapabilities_POST(url=WFS_get_url(),version='1.1.0')
WFS_util_showxml(cap2,lines=2:3)
}
}
