% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_getfeature.R
\name{WFS_getfeature}
\alias{WFS_getfeature}
\title{Retrieve WFS information with the GetFeature request}
\usage{
WFS_getfeature(
  typename,
  ...,
  url = WFS_get_url(),
  version = WFS_get_version(),
  debug = F
)
}
\arguments{
\item{typename}{Character with the name of a feature (such as e.g. found by using \code{\link[=WFS_featuretypes]{WFS_featuretypes()}} )}

\item{...}{optional arguments for the \code{GetFeature} request. See \strong{Details}}

\item{url}{URL of the WFS service. See \code{\link[=WFS_get_url]{WFS_get_url()}} for the default}

\item{version}{software version for WFS service request. See \code{\link[=WFS_get_version]{WFS_get_version()}} for the default#'}

\item{debug}{Logical indicating only json output required}
}
\value{
a \code{sf} object with the requested information (the 'id' variable and the geometry will always be included) or a character string with an error message
}
\description{
Retrieve the requested information
}
\details{
Arguments that are recognized by the GetFeature request are:
\itemize{
\item \code{bbox        } (not in combination with \code{cql_filter} or \code{filter})
\item \code{cql_filter  }
\item \code{filter      }
\item \code{srsname     } indicate the crs for the coordinates e.g. \code{srsname='EPSG:4326'}
\item \code{propertyname} the name of the fields to retrieve. The \code{id} and \code{geometry} fields will always be included.
\item \code{startIndex  } number of features to skip before retrieving features ( the output with \code{startindex=1} will start with the second feature )
\item \code{maxfeatures } (only for version \verb{1.1.0}) or \code{count} (only for version \verb{2.0.0}) indicates the number of features to retrieve
}
}
\examples{
\dontrun{
typename <- 'topp:gidw_groenbomen'
wfs1 = WFS_getfeature(typename)
wfs2 = WFS_getfeature(typename,
   cql_filter= r"(boom_omschrijf='Prunus serrulata ''Kanzan''')" ) # double internal quotes !
wfs3 = WFS_getfeature(typename,
   bbox = '119038,479244,119500,479500') # in EPSG:28992  (default see WFS_featuretypes)
wfs3a = WFS_getfeature(typename,
  cql_filter= r"( bbox(geometrie,119038,479244,119500,479500,'EPSG:28992') and (boom_omschrijf='Acer freemanii ''Elegant'''))")
wfs3b = WFS_getfeature(typename,
  cql_filter= r"( bbox(geometrie,119038,479244,119500,479500,'EPSG:28992'))")
wfs4 = WFS_getfeature(typename,
   bbox = '119038,479244,119500,479500,EPSG:28992',  # in EPSG:28992
   srsname = 'EPSG:4326') # but request coordinates in  WGS84 (EPSG:4326)
f5 = fg("Filter"
, fg(
  "PropertyIsEqualTo"
  , bg('PropertyName', 'topp:boom_omschrijf')
  , bg('Literal',  r"(Acer freemanii 'Elegant')" )#'
  )
)
f5
wfs5 = WFS_getfeature(typename,
   srsname = 'EPSG:4326', # but request coordinates in  WGS84 (EPSG4326)
   filter  = f5)
}
}
