% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_getfeature2.R
\name{WFS_POST_request}
\alias{WFS_POST_request}
\title{Retrieve information with the POST request}
\usage{
WFS_POST_request(
  request,
  url = WFS_get_url(),
  debug = F,
  to_sf = T,
  sfverbose = F,
  httrverbose = rep(F, 4)
)
}
\arguments{
\item{request}{Character string with POST request}

\item{url}{URL of the WFS service. See \code{\link[=WFS_get_url]{WFS_get_url()}} for the default}

\item{debug}{Logical indicating the httr response is to be returned}

\item{sfverbose}{Logical indicating if \code{\link[sf:st_read]{sf::read_sf()}} messages will be displayed}

\item{httrverbose}{Logical vector of up to four entries to be used in \code{\link[httr:verbose]{httr::verbose()}}}
}
\value{
a \code{json} object when that is returned but converted to an \code{sf} object when \code{to_sf==T}.
When the GET returns an \code{xml} object it is returned.
}
\description{
Retrieve the requested information
}
\examples{
\dontrun{
}
}
