% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_utils.R
\name{WFS_util_attrs}
\alias{WFS_util_attrs}
\title{Retrieves the (xmlns) attributes from a GetCapabilities document}
\usage{
WFS_util_attrs(cap, skip = c("none", "known", "version"), as.text = T)
}
\arguments{
\item{cap}{XML document containg the GetCapabilities information}

\item{skip}{Character string which of the attributes will be removed. See Details}

\item{as.text}{Logical scalar. If \code{TRUE} indicates that the result will be a character string
(ready for inclusion in a XML request). Otherwise the result is a named list.}
}
\value{
Character string or list depending on the argument \code{as.text}
}
\description{
Retrieves the (xmlns) attributes from a GetCapabilities document
}
\details{
The user can indicate with the \code{skip} argument which attributes will be left out:
\itemize{
\item \code{none} : all will be returned
\item \code{known} : all the known (standard) attributes will be left out:
\code{version}, \code{schemaLocation}, \code{updateSequence}, \code{xmlns:wfs},
\code{xmlns:xsi}, \code{xmlns}, \code{xmlns:ows}, \code{xmlns:gml}, \code{xmlns:ogc},
\code{xmlns:fes}, \code{xmlns:xlink} and \code{xmlns:xs}
\item \code{version}: only the version attribute will be left out
}
}
\examples{
\dontrun{
base_url  <- "https://geodata.nationaalgeoregister.nl/wijkenbuurten2019/wfs"
cap1 = WFS_getcapabilities_POST(base_url,'1.1.0')
WFS_util_attrs(cap1,skip='version',as.text=T)
}
}
