% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_utils.R
\name{WFS_util_keep_unique}
\alias{WFS_util_keep_unique}
\title{Keep first or last entry of list entries with duplicated names}
\usage{
WFS_util_keep_unique(mylist, keep_first = T, ignore.case = T)
}
\arguments{
\item{mylist}{A named list}

\item{ignore.case}{Logical scalar indicating if case is ignored when comparing for duplicate names}

\item{keep.first}{Logical scalar indicating if the first (when \code{TRUE}) or last entry (when \code{FALSE}) with duplicate names will kept}
}
\value{
A copy of \code{mylist} with duplicate names removed (apart from the first when \code{keep_first=TRUE} or last otherwise )
}
\description{
Keep first or last entry of list entries with duplicated names
}
\examples{
L <- list( a = 1, A = 2, aA = 3, Aa = 4)
WFS_util_keep_unique(L,keep_first = T, ignore.case=T)
# list(a = 1, aA = 3)
}
