% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_utils.R
\name{WFS_util_xmlns_defs}
\alias{WFS_util_xmlns_defs}
\title{Provides the standard xmlns definitions}
\usage{
WFS_util_xmlns_defs(version = WFS_get_version(), as.txt = T)
}
\arguments{
\item{version}{software version for WFS service request. See \code{\link[=WFS_get_version]{WFS_get_version()}} for the default}

\item{as.txt}{Logical indicating if character string is to be returned (T) or a list (F)}
}
\value{
character string or list with the xmlns definitions
}
\description{
Provides the standard xmlns definitions
}
\examples{
\dontrun{
WFS_util_xmlns_defs(as.txt=T)
}
}
