% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_convert.R
\name{wfsconvert}
\alias{wfsconvert}
\alias{convert_bbox}
\alias{convert_points}
\title{Convert coordinates}
\usage{
convert_bbox(coords, crs_in, crs_out)

convert_points(coords, crs_in, crs_out, out_matrix = T, keep_names = T)
}
\arguments{
\item{coords}{Numeric vector with four elements indicating the bounding box (for \code{convert_bbox} ) or a numeric vector of even length
or numeric matrix with two columns (for \code{convert_points} )}

\item{crs_in}{Character string with the name of the input crs (e.g. \code{"EPSG:4326"} )}

\item{crs_out}{Character string with the name of the output crs (e.g. \code{"EPSG:28992"} )}

\item{out_matrix}{Logical indicating that the output of  \code{convert_points} should be matrix}

\item{keep_names}{Logical indicating that the output of \code{convert_points} keeps the column names when the input is a matrix}
}
\value{
The \code{convert_bbox} returns a numeric vector and the  \code{convert_points} returns a numeric matrix with two columns
}
\description{
\itemize{
\item The function \code{convert_bbox} converts a bbox from one CRS to another
\item The function \code{convert_points} converts a set of points (in vector or matrix form) from one CRS to another
}
}
\examples{
bbox_28992 <- c(119103, 480726, 119160, 481078)
bbox_4326  <- convert_bbox(bbox_28992,"EPSG:28992","EPSG:4326")
points_4326 <- convert_points(bbox_28992,"EPSG:28992","EPSG:4326")
}
