% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_init.R
\name{WFS_set_url}
\alias{WFS_set_url}
\alias{WFS_get_url}
\alias{WFS_set_version}
\alias{WFS_get_version}
\title{Get/Set url and request version for the WFS service}
\usage{
WFS_set_url(url = NULL)

WFS_get_url()

WFS_set_version(version = NULL)

WFS_get_version()
}
\arguments{
\item{url}{NULL for the default url ( \strong{https://geoweb.amstelveen.nl/geoserver/topp/wfs} ) or the url of the WFS service otherwise}

\item{version}{NULL for the default ( \strong{1.1.0} ) request version or an alternative version (e.g. 2.0.0 ) otherwise}
}
\value{
Character vector (invisible for the set version) with the url or version to be used
}
\description{
Determines from which WFS service information will be extracted and the version of the WFS requests to use. If the url is not specified before
the default url (https://geoweb.amstelveen.nl/geoserver/topp/wfs) will be used. If no version is specified the default version (1.1.0) is be used.
\cr\cr
Note that these values are used as the default values in the various request functions. They can always be overwritten in these functions.
}
\examples{
WFS_set_url("https://geoweb.amstelveen.nl/geoserver/topp/wfs")
WFS_get_url()
WFS_set_version()
WFS_set_version('2.0.0')
WFS_get_version()
}
