% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_tags.R
\name{fg}
\alias{fg}
\alias{bg}
\title{Tag functions fg and bg}
\usage{
fg(tag, ..., ta = NULL, sep = WFS_get_sep())

bg(tag, value = "")
}
\arguments{
\item{tag}{Character with tag that will be used as both start and end tag}

\item{...}{Character arguments that will be inserted between start and end tag}

\item{ta}{Character with additional information for start tag}

\item{sep}{NULL for the default separator (set by \code{\link[=WFS_set_sep]{WFS_set_sep()}}) or required separator otherwise}

\item{value}{Character argument that will be inserted between start and end tag}
}
\value{
Character vector with the generated xml element
}
\description{
The \code{fg} and \code{bg} functions are auxiliary functions to create xml elements.
Using these functions only the start tag and contents have to be provided; \cr
e.g.
\cr\cr \code{bg('PropertyName', 'topp:objec_omschrijf')}
\cr  will create
\cr \verb{<PropertyName>topp:objec_omschrijf</PropertyName>}
\cr\cr The function \code{fg} allows additional arguments for the tag and allows subtags.
See \strong{Examples} and the functions in \link{wfsfilteraux}
}
\examples{
fg(
  'PropertyIsLike'
  , bg('PropertyName', 'topp:objec_omschrijf')
  , bg('Literal', '.unmoolen')
  , ta = 'wildCard="*" singleChar="." escape="!"'
  )
}
