% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_getcapabilities.R
\name{WFS_getcapabilities}
\alias{WFS_getcapabilities}
\title{GetCapabilities for the WFS service}
\usage{
WFS_getcapabilities(
  url = WFS_get_url(),
  version = WFS_get_version(),
  debug = F,
  verbose = F,
  out_path = NULL
)
}
\arguments{
\item{url}{URL of the WFS service. See \code{\link[=WFS_get_url]{WFS_get_url()}} for the default}

\item{version}{software version for WFS service request. See \code{\link[=WFS_get_version]{WFS_get_version()}} for the default}

\item{debug}{Logical indicating the httr response is to be returned}

\item{verbose}{Logical indicating full request and httr response code will be displayed}

\item{out_path}{(optional) path where the xml result is to be saved}
}
\value{
xml document with the \code{GetCapabilities} information for this WFS service
or character string 'UNEXPECTED ERROR' when an error was encountered
}
\description{
Retrieve the GetCapabilities document with request version \code{version} for WFS service \code{url}
}
\examples{
\dontrun{
WFS_getcapabilities("https://geoweb.amstelveen.nl/geoserver/topp/wfs","1.1.0",out_path="gc.xml")
}
}
