% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_getfeature.R
\name{WFS_getfeature}
\alias{WFS_getfeature}
\title{Retrieve WFS information with the GetFeature request}
\usage{
WFS_getfeature(
  typename,
  ...,
  httrType = c("GET", "POST"),
  url = WFS_get_url(),
  version = WFS_get_version(),
  debug = F,
  sfverbose = F,
  httrverbose = rep(F, 4)
)
}
\arguments{
\item{typename}{Character with the name of a feature (such as e.g. found by using \code{\link[=WFS_featuretypes]{WFS_featuretypes()}} )}

\item{...}{optional arguments for the \code{GetFeature} request. See \strong{Details}}

\item{httrType}{Character with the type of httr request: \code{GET} or \code{POST}}

\item{url}{URL of the WFS service. See \code{\link[=WFS_get_url]{WFS_get_url()}} for the default}

\item{version}{software version for WFS service request. See \code{\link[=WFS_get_version]{WFS_get_version()}} for the default}

\item{debug}{Logical indicating the httr response is to be returned}

\item{sfverbose}{Logical indicating if \code{\link[sf:st_read]{sf::read_sf()}} messages will be displayed}

\item{httrverbose}{Logical vector of up to four entries to be used in \code{\link[httr:verbose]{httr::verbose()}}}
}
\value{
a \code{sf} object with the requested information (the 'id' variable and the geometry will always be included)
or a character string with an error message
}
\description{
Retrieve the requested information.
}
\details{
Arguments that are recognized by the GetFeature request are:
\itemize{
\item \code{bbox        } (not in combination with \code{cql_filter} or \code{filter})
\item \code{cql_filter  }
\item \code{filter      }
\item \code{resultType} - default 'results' and alternative 'hits'. In the latter case only the number of matched features (\code{numbeOfFeatures} in 1.1.0 or \code{numberMatched} in 2.0.0) is returned
\item \code{srsname     } indicate the crs for the coordinates of the output e.g. \code{srsname='EPSG:4326'}
\item \code{propertyname} the name of the fields to retrieve. The \code{id} and \code{geometry} fields will always be included.
\item \code{startIndex  } number of features to skip before retrieving features ( the output with \code{startindex=1} will start with the second feature )
\item \code{maxfeatures } (version \verb{1.1.0}) or \code{count} (version \verb{2.0.0}) indicates the number of features to retrieve. When the wrong argument is specified it will be translated in the other.
The \code{GetFeature} argument \code{outputFormat} has value 'application/json' but this can be overwritten
}
}
\examples{
\dontrun{

}
}
