% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_getfeature2.R
\name{WFS_POST_request}
\alias{WFS_POST_request}
\title{Retrieve information with the POST request}
\usage{
WFS_POST_request(
  request,
  url = WFS_get_url(),
  debug = F,
  to_sf = T,
  sfverbose = F,
  WFSverbose = rep(F, 4)
)
}
\arguments{
\item{request}{Character string with expanded url}

\item{debug}{Logical indicating the httr response is to be returned}

\item{verbose}{Logical indicating full request and httr response code will be displayed}
}
\value{
a \code{json} object when that is returned but converted to an \code{sf} object when \code{to_sf==T}.
When the GET returns an \code{xml} object it is returned but unpacked if it is an \code{ExceptionReport}.
}
\description{
Retrieve the requested information
}
\examples{
\dontrun{
}
}
