% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_utils.R
\name{WFS_util_parameter_values}
\alias{WFS_util_parameter_values}
\title{Retrieves the allowed values for the parameters of WFS operations}
\usage{
WFS_util_parameter_values(cap, operation = "??", parameter = "??")
}
\arguments{
\item{cap}{XML document containg the GetCapabilities information}

\item{operation}{Character string indicating one of the WFS operations}

\item{parameter}{Character string indicating one of the parameters of the operation}
}
\value{
Character vector with the allowed values for the parameter
}
\description{
Retrieves the allowed values for the parameters of WFS operations
}
\examples{
\dontrun{
WFS_util_parameter_values(cap,operation='GetFeature',parameter='outputFormat')
WFS_util_parameter_values(cap,operation='DescribeFeatureType',parameter='outputFormat')
}
}
