% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_utils.R
\name{WFS_util_showxml}
\alias{WFS_util_showxml}
\title{Show XML fragment on screen}
\usage{
WFS_util_showxml(
  myxml,
  width = 2000,
  max_n = 20,
  lines = NULL,
  screenwidth = getOption("width", 100)
)
}
\arguments{
\item{myxml}{XML document, node_set or node}

\item{width}{Numeric scalar passed to the xml2 print function indicating how many characters of the xml lines will be shown}

\item{lines}{Numerical vector indicating which lines of the print output will actually be shown. \code{NULL} indicates all. This argument is passed to \code{\link[HOQCutil:cap.out]{HOQCutil::cap.out()}}}

\item{screenwidth}{Numeric scalar scalar indicating at which point the print output will be wrapped. This argument is passed to \code{\link[HOQCutil:cap.out]{HOQCutil::cap.out()}} as its 'with' argument}

\item{max_n=20}{Numeric scalar passed to the xml2 print function indicating how many  xml lines will be printed}

\item{verbose}{Logical indicating full request and httr response code will be displayed}
}
\value{
NULL (invisible)
}
\description{
This is a cover function for the XML  \code{xml2:::print.xml_document}, \code{print.xml_nodeset} and \code{print.xml_node} print functions
}
\examples{
\dontrun{
WFS_util_showxml(cap1,lines=2:3)
}
}
