% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_utils.R
\name{WFS_util_unify_names}
\alias{WFS_util_unify_names}
\title{replace argument names by their uniform counterparts}
\usage{
WFS_util_unify_names(mylist, vnames)
}
\arguments{
\item{mylist}{A named list}

\item{vnames}{A list of three character vectors:
\itemize{
\item uniform names
\item first set of names
\item second set of names
}}
}
\value{
A copy of \code{mylist} with names in both sets replaced by the corresponding uniform one
}
\description{
replace argument names by their uniform counterparts
}
