% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_utils.R
\name{WFS_util_v12_names}
\alias{WFS_util_v12_names}
\title{incompatible argument names in versions \verb{1.1.0} and \verb{2.0.0}}
\usage{
WFS_util_v12_names()
}
\value{
A list of three character vectors:
\itemize{
\item uniform names of temporary use
\item corresponding version \verb{1.1.0} names
\item corresponding version \verb{2.0.0} names
}
}
\description{
incompatible argument names in versions \verb{1.1.0} and \verb{2.0.0}
}
