% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFS_filter_aux.R
\name{wfsfilteraux}
\alias{wfsfilteraux}
\alias{build_filter}
\alias{propertyname_xml}
\alias{propeq_xml}
\alias{bbox_xml}
\title{Auxiliary functions to create filters}
\usage{
build_filter(..., version = WFS_get_version(), sep = WFS_get_sep())

propertyname_xml(propname, version, nopref = T)

propeq_xml(propname, propvalue, version = WFS_get_version())

bbox_xml(gemprop, crs_in, coords, version = WFS_get_version())
}
\arguments{
\item{...}{XML elements to be added to filter}

\item{version}{Character string with the WFS request version}

\item{sep}{NULL for the default separator (set by \code{\link[=WFS_set_sep]{WFS_set_sep()}}) or required separator otherwise}

\item{propname}{Character string with the \code{PropertyName} (1.1.0) or \code{ValueReference} (2.0.0)}

\item{nopref}{Logical scalar indicating the 'ogc:' pref is not to be used in \code{propertyname_xml}}

\item{propvalue}{Character string to filter with}

\item{gemprop}{Character string with the name of the geometric field}

\item{crs_in}{Character string with the name of the input crs (e.g. \code{"EPSG:4326"} )}

\item{coords}{Numeric vector with four elements indicating the bounding box}
}
\value{
Character vector with the created filter or xml fragment.
}
\description{
\itemize{
\item The function \code{build_filter} creates a filter for the indicated version by inserting the appropriate 'xmlns' urls
in the Filter tag and inserting the XML elements from the \code{...} argument i.e. all arguments except \code{version} and \code{sep}.
While evaluating these elements the version and separator are temporarily set to the arguments \code{version} and \code{sep} of
\code{build_filter} unless explicitly overwritten.
\item The function \code{propeq_xml} creates a \code{PropertyIsEqualTo} XML clause
\item The function \code{propertyname_xml} creates a \code{PropertyName} or \code{ValueReference} XML clause
\item The function \code{bbox_xml} creates a \code{BBOX} XML clause
}
}
\examples{
bbox_28992 <- c(119103, 480726, 119160, 481078)
f1 <- build_filter(version='1.0.0',
   fg("And"
       , propeq_xml('topp:boom_omschrijf',"Alnus glutinosa 'Laciniata'")
       , bbox_xml("geometrie","EPSG:28992",bbox_28992)
     )
  )
bbox_4326 <- convert_bbox(bbox_28992,"EPSG:28992","EPSG:4326")
f2 <- build_filter(version='2.0.0',
   fg("And"
       , propeq_xml('topp:boom_omschrijf',"Alnus glutinosa 'Laciniata'")
       , bbox_xml("geometrie","EPSG:4326",bbox_4326)
     )
  )
pn_construct <- propertyname_xml('han','2.0.0',nopref =F)
bbox_28992 <- c(119103, 480726, 119160, 481078)
bbox_4326  <- convert_bbox(bbox_28992,"EPSG:28992","EPSG:4326")
points_4326 <- convert_points(bbox_28992,"EPSG:28992","EPSG:4326")
}
